shorten execution path
