+++ src/core/timeout/TimeoutFallbackHandler.ts
@@ -1,15 +1,32 @@
 import type { ToolName } from "@roo-code/types"
-import { formatResponse } from "../prompts/responses"
-import type { TimeoutFallbackContext } from "../prompts/instructions/timeout-fallback"
-import { TimeoutFallbackGenerator, type TimeoutFallbackResult } from "./TimeoutFallbackGenerator"
 import type { Task } from "../task/Task"
+import type { SingleCompletionHandler } from "../../api"
+import {
+	createTimeoutFallbackPrompt,
+	parseTimeoutFallbackResponse,
+	type TimeoutFallbackContext,
+} from "../prompts/instructions/timeout-fallback"
+import { formatResponse } from "../prompts/responses"
+
+export interface TimeoutFallbackResult {
+	success: boolean
+	toolCall?: {
+		name: "ask_followup_question"
+		params: {
+			question: string
+			follow_up: string
+		}
+	}
+	error?: string
+}
 
 /**
- * Generates AI-powered fallback suggestions for timeout scenarios
+ * Unified timeout fallback handler that generates AI-powered fallback suggestions
+ * and creates timeout responses in a single optimized flow
  */
 export class TimeoutFallbackHandler {
 	/**
-	 * Create a timeout response with AI-generated fallback question
+	 * Create a timeout response with AI-generated fallback question in a single optimized query
 	 */
 	public static async createTimeoutResponse(
 		toolName: ToolName,
@@ -40,12 +57,11 @@ export class TimeoutFallbackHandler {
 				: undefined,
 		}
 
-		// Generate AI-powered fallback (with static fallback if AI fails)
-		const aiResult = await TimeoutFallbackGenerator.generateAiFallback(aiContext, task)
+		// Generate AI-powered fallback (with static fallback if AI fails) in a single call
+		const aiResult = await this.generateAiFallback(aiContext, task)
 
 		if (aiResult.success && aiResult.toolCall) {
-			// Instead of injecting the tool call, we'll return a response that instructs
-			// the model to ask a follow-up question in its next message
+			// Return a response that instructs the model to ask a follow-up question
 			const { question, follow_up } = aiResult.toolCall.params
 
 			// Format the response to explicitly instruct the model to ask the follow-up question
@@ -65,5 +81,107 @@ This is required to help the user decide how to proceed after the timeout.`
 
 		// This should rarely happen since generateAiFallback always provides static fallback
 		return `${baseResponse}\n\nThe operation timed out. Please consider breaking this into smaller steps or trying a different approach.`
+	}
+
+	/**
+	 * Generate an AI-powered ask_followup_question tool call for timeout scenarios
+	 */
+	public static async generateAiFallback(
+		context: TimeoutFallbackContext,
+		task?: Task,
+	): Promise<TimeoutFallbackResult> {
+		// Try to use AI to generate contextual suggestions
+		if (task?.api && "completePrompt" in task.api) {
+			try {
+				const aiResult = await this.generateAiSuggestions(context, task.api as SingleCompletionHandler)
+				if (aiResult.success) {
+					return aiResult
+				}
+			} catch (error) {
+				// AI failed, fall through to static suggestions
+			}
+		}
+
+		// Fallback to static suggestions if AI fails or is unavailable
+		const toolCall = this.generateStaticToolCall(context)
+
+		return {
+			success: true,
+			toolCall,
+		}
+	}
+
+	/**
+	 * Generate AI-powered suggestions using the task's API handler
+	 */
+	private static async generateAiSuggestions(
+		context: TimeoutFallbackContext,
+		apiHandler: SingleCompletionHandler,
+	): Promise<TimeoutFallbackResult> {
+		try {
+			const prompt = createTimeoutFallbackPrompt(context)
+			const aiResponse = await apiHandler.completePrompt(prompt)
+
+			// Parse the AI response to extract suggestions
+			const suggestions = parseTimeoutFallbackResponse(aiResponse)
+
+			if (suggestions.length === 0) {
+				throw new Error("No valid suggestions generated by AI")
+			}
+
+			const question = `The ${context.toolName} operation timed out after ${Math.round(context.timeoutMs / 1000)} seconds. How would you like to proceed?`
+
+			const followUpXml = suggestions
+				.map((suggestion) =>
+					suggestion.mode
+						? `<suggest mode="${suggestion.mode}">${suggestion.text}</suggest>`
+						: `<suggest>${suggestion.text}</suggest>`,
+				)
+				.join("\n")
+
+			return {
+				success: true,
+				toolCall: {
+					name: "ask_followup_question",
+					params: {
+						question,
+						follow_up: followUpXml,
+					},
+				},
+			}
+		} catch (error) {
+			return {
+				success: false,
+				error: error instanceof Error ? error.message : "Unknown error generating AI suggestions",
+			}
+		}
+	}
+
+	/**
+	 * Generate static fallback suggestions when AI is unavailable
+	 */
+	private static generateStaticToolCall(context: TimeoutFallbackContext): TimeoutFallbackResult["toolCall"] {
+		const suggestions = formatResponse.timeoutFallbackSuggestions.generateContextualSuggestions(
+			context.toolName,
+			context.toolParams,
+		)
+
+		const question = `The ${context.toolName} operation timed out after ${Math.round(context.timeoutMs / 1000)} seconds. How would you like to proceed?`
+
+		const followUpXml = suggestions
+			.map((suggestion) =>
+				suggestion.mode
+					? `<suggest mode="${suggestion.mode}">${suggestion.text}</suggest>`
+					: `<suggest>${suggestion.text}</suggest>`,
+			)
+			.join("\n")
+
+		return {
+			name: "ask_followup_question",
+			params: {
+				question,
+				follow_up: followUpXml,
+			},
+		}
 	}
 }
