+++ src/core/timeout/index.ts
@@ -1,9 +1,8 @@
 export { TimeoutManager, timeoutManager } from "./TimeoutManager"
 export { ToolExecutionWrapper } from "./ToolExecutionWrapper"
 export { TimeoutFallbackHandler } from "./TimeoutFallbackHandler"
-export { TimeoutFallbackGenerator } from "./TimeoutFallbackGenerator"
 
 export type { TimeoutConfig, TimeoutResult, TimeoutEvent } from "./TimeoutManager"
 export type { ToolExecutionOptions } from "./ToolExecutionWrapper"
-export type { TimeoutFallbackResult } from "./TimeoutFallbackGenerator"
+export type { TimeoutFallbackResult } from "./TimeoutFallbackHandler"
 export type { TimeoutFallbackContext } from "../prompts/instructions/timeout-fallback"
